Injection prototype and plotting updates
+ working Injection prototype
+ diag_spectra() mostly updated
