+++ run_blip.py
@@ -3,9 +3,10 @@ import numpy as np
 import sys, configparser, subprocess
 from src.makeLISAdata import LISAdata
 from src.likelihoods import likelihoods
-from src.models import Model
+from src.models import Model, Injection
 from tools.plotmaker import plotmaker
 from tools.plotmaker import mapmaker
+import matplotlib
 import matplotlib.pyplot as plt
 from astropy import units as u
 from multiprocessing import Pool
@@ -30,10 +31,11 @@ class LISA(LISAdata, Model):
         LISAdata.__init__(self, params, inj)
 
         # Make noise spectra
-        self.which_noise_spectrum()
+#        self.which_noise_spectrum(params,inj,self.fdata,self.f0,self.tsegmid)
 
-        if self.inj['doInj']:
-            self.which_astro_signal()
+#        if self.inj['doInj']:
+#            self.Injection = Injection()
+#            self.which_astro_signal()
 
         # Generate or get mldc data
         if self.params['mldc']:
@@ -61,35 +63,66 @@ class LISA(LISAdata, Model):
         to generate data. Return Frequency domain data.
         '''
 
+        ## define the splice segment duration
+        tsplice = 1e4
+        ## the segments to be splices are half-overlapping
+        nsplice = 2*int(self.params['dur']/tsplice) + 1
+        ## arrays of segmnent start and mid times
+        tsegmid = self.params['tstart'] +  (tsplice/2.0) * np.arange(nsplice) + (tsplice/2.0)
+        ## Number of time-domain points in a splice segment
+        Npersplice = int(self.params['fs']*tsplice)
+        ## leave out f = 0
+        frange = np.fft.rfftfreq(Npersplice, 1.0/self.params['fs'])[1:]
+        ## the charecteristic frequency of LISA, and the scaled frequency array
+        fstar = 3e8/(2*np.pi*self.armlength)
+        f0 = frange/(2*fstar)
+        
+        ## Build the Injection object
+        self.Injection = Injection(self.params,self.inj,frange,f0,tsegmid)
+        
+        ## assign a couple additional universal injection attributes needed in add_sgwb_data()
+        self.Injection.Npersplice = Npersplice
+        self.Injection.nsplice = nsplice
+        
         # Generate TDI noise
-        times, self.h1, self.h2, self.h3 = self.gen_noise_spectrum()
+        times, self.h1, self.h2, self.h3 = self.Injection.components['noise'].gen_noise_spectrum()
         delt = times[1] - times[0]
 
         # Cut to required size
         N = int((self.params['dur'])/delt)
         self.h1, self.h2, self.h3 = self.h1[0:N], self.h2[0:N], self.h3[0:N]
 
-        # Generate TDI isotropic signal
-        if self.inj['doInj']:
-            if self.inj['injtype'] == 'multi':
-                h1_gw_i, h2_gw_i, h3_gw_i, times = self.add_sgwb_data(multi='i')
-                h1_gw_a, h2_gw_a, h3_gw_a, times = self.add_sgwb_data(multi='a')
-                h1_gw_i, h2_gw_i, h3_gw_i = h1_gw_i[0:N], h2_gw_i[0:N], h3_gw_i[0:N]
-                h1_gw_a, h2_gw_a, h3_gw_a = h1_gw_a[0:N], h2_gw_a[0:N], h3_gw_a[0:N]
+        ## create time-domain contribution from each injection component that isn't noise
+        for component in self.Injection.sgwb_component_names:
+            h1_gw, h2_gw, h3_gw, times = self.add_sgwb_data(self.Injection.components[component])
+            h1_gw, h2_gw, h3_gw = h1_gw[0:N], h2_gw[0:N], h3_gw[0:N]
 
-                # Add gravitational-wave time series to noise time-series
-                self.h1 = self.h1 + h1_gw_i + h1_gw_a
-                self.h2 = self.h2 + h2_gw_i + h2_gw_a
-                self.h3 = self.h3 + h3_gw_i + h3_gw_a
-            else:
-                h1_gw, h2_gw, h3_gw, times = self.add_sgwb_data()
-
-                h1_gw, h2_gw, h3_gw = h1_gw[0:N], h2_gw[0:N], h3_gw[0:N]
-    
-                # Add gravitational-wave time series to noise time-series
-                self.h1 = self.h1 + h1_gw
-                self.h2 = self.h2 + h2_gw
-                self.h3 = self.h3 + h3_gw
+            # Add gravitational-wave time series to noise time-series
+            self.h1 = self.h1 + h1_gw
+            self.h2 = self.h2 + h2_gw
+            self.h3 = self.h3 + h3_gw
+        
+        
+#        if self.inj['injtype'] == 'multi':
+#            h1_gw_i, h2_gw_i, h3_gw_i, times = self.add_sgwb_data(multi='i')
+#            h1_gw_a, h2_gw_a, h3_gw_a, times = self.add_sgwb_data(multi='a')
+#            h1_gw_i, h2_gw_i, h3_gw_i = h1_gw_i[0:N], h2_gw_i[0:N], h3_gw_i[0:N]
+#            h1_gw_a, h2_gw_a, h3_gw_a = h1_gw_a[0:N], h2_gw_a[0:N], h3_gw_a[0:N]
+#
+#            # Add gravitational-wave time series to noise time-series
+#            self.h1 = self.h1 + h1_gw_i + h1_gw_a
+#            self.h2 = self.h2 + h2_gw_i + h2_gw_a
+#            self.h3 = self.h3 + h3_gw_i + h3_gw_a
+#        else:
+#            for injection_component in 
+#            h1_gw, h2_gw, h3_gw, times = self.add_sgwb_data()
+#
+#            h1_gw, h2_gw, h3_gw = h1_gw[0:N], h2_gw[0:N], h3_gw[0:N]
+#
+#            # Add gravitational-wave time series to noise time-series
+#            self.h1 = self.h1 + h1_gw
+#            self.h2 = self.h2 + h2_gw
+#            self.h3 = self.h3 + h3_gw
 
         self.timearray = times[0:N]
         if delt != (times[1] - times[0]):
@@ -268,133 +301,143 @@ class LISA(LISAdata, Model):
         Np, Na = 10**self.inj['log_Np'], 10**self.inj['log_Na']
 
         # Modelled Noise PSD
-        C_noise = self.instr_noise_spectrum(self.fdata,self.f0, Np, Na)
+        C_noise = self.Injection.components['noise'].instr_noise_spectrum(self.fdata,self.f0, Np, Na)
 
         # Extract noise auto-power
         S1, S2, S3 = C_noise[0, 0, :], C_noise[1, 1, :], C_noise[2, 2, :]
+        
+        ## this is kinda hacky for now
+        ## set up a color-coder for the different models somewhere later
+#        matplotlib.rcParams['axes.prop_cycle'] = matplotlib.cycler(color=["mediumorchid", "teal", "goldenrod","slategray","navy"])
+        
+        plt.close()
+        for component_name in self.Injection.sgwb_component_names:
+            S1_gw = self.Injection.plot_injected_spectra(component_name,fs_new=self.fdata,convolved=True,legend=True,channel='1',return_PSD=True,lw=0.75)
+            S2_gw, S3_gw = self.Injection.compute_convolved_spectra(component_name,fs_new=self.fdata,channel='2'), self.Injection.compute_convolved_spectra(component_name,fs_new=self.fdata,channel='3')
+            S1, S2, S3 = S1+S1_gw, S2+S2_gw, S3+S3_gw
+#        
+#        if self.inj['injtype'] != 'noise_only':
+#            if self.inj['injtype'] == 'sph_sgwb' or self.inj['injtype']=='astro':
+#                ## Response matrix : shape (3 x 3 x freq x time) if isotropic
+#                ## set up different use cases
+#                if self.inj['injtype'] == 'astro':
+#                    if self.inj['injbasis'] == 'sph':
+#                        summ_response_mat = np.sum(self.add_astro_signal(self.f0,self.tsegmid)*self.alms_inj[None,None,None,None,:],axis=-1)
+#                    elif self.inj['injbasis'] == 'sph_lmax':
+#                        summ_response_mat = np.sum(self.add_astro_signal(self.f0,self.tsegmid,set_almax=self.inj_almax)*self.alms_inj[None,None,None,None,:],axis=-1)
+#                    elif self.inj['injbasis'] == 'pixel':
+#                        summ_response_mat = self.add_astro_signal(self.f0,self.tsegmid,self.skymap_inj)  
+#                else:
+#                    summ_response_mat = np.sum(self.response_mat*self.alms_inj[None, None, None, None, :], axis=-1)
+#                
+#                # extra auto-power GW responses
+#                R1 = np.real(summ_response_mat[0, 0, :, :])
+#                R2 = np.real(summ_response_mat[1, 1, :, :])
+#                R3 = np.real(summ_response_mat[2, 2, :, :])
+#            
+#            elif self.inj['injtype'] == 'multi':
+#                if self.inj['injbasis'] == 'sph':
+#                    summ_response_mat_a = np.sum(self.add_astro_signal_a(self.f0,self.tsegmid)*self.alms_inj[None,None,None,None,:],axis=-1)
+#                    response_mat_i = self.add_astro_signal_i(self.f0, self.tsegmid)
+#                    R1_a = np.real(summ_response_mat_a[0, 0, :, :])
+#                    R2_a = np.real(summ_response_mat_a[1, 1, :, :])
+#                    R3_a = np.real(summ_response_mat_a[2, 2, :, :])
+#                    R1_i = np.real(response_mat_i[0, 0, :, :])
+#                    R2_i = np.real(response_mat_i[1, 1, :, :])
+#                    R3_i = np.real(response_mat_i[2, 2, :, :])
+#                else:
+#                    raise ValueError("Multi-SGWB injections only supports sph basis injections for now.")
+#            else:
+#                # extra auto-power GW responses
+#                R1 = np.real(self.add_astro_signal(self.f0, self.tsegmid)[0, 0, :, :])
+#                R2 = np.real(self.add_astro_signal(self.f0, self.tsegmid)[1, 1, :, :])
+#                R3 = np.real(self.add_astro_signal(self.f0, self.tsegmid)[2, 2, :, :])
+#            
+#            
+#            # Hubble constant
+#            H0 = 2.2*10**(-18)
+#            if self.inj['injtype'] == 'multi':
+#                Omegaf_i = (10**self.inj['log_omega0_i'])*(self.fdata/self.params['fref'])**self.inj['alpha_i']
+#                Omegaf_a = (10**self.inj['log_omega0_a']) * (self.fdata/self.params['fref'])**self.inj['alpha_a'] \
+#                                * 0.5 * (1+np.tanh((self.inj['f_cut_a']-self.fdata)/self.inj['f_scale_a']))
+#                Sgw_a = (3.0*(H0**2)*Omegaf_a)/(4*np.pi*np.pi*self.fdata**3)
+#                Sgw_i = (3.0*(H0**2)*Omegaf_i)/(4*np.pi*np.pi*self.fdata**3)
+#            else:
+#                # Calculate astrophysical power law noise
+#                if self.inj['spectral_inj'] == 'population':
+#                    ## population-derived power spectra
+#                    Sgw = self.pop2spec(self.inj['popfile'],self.fdata,self.params['dur']*u.s,names=self.inj['columns'],sep=self.inj['delimiter'])*4 ##h^2 = 1/2S_A = 1/2 * 1/2S_GW
+#                    # Power spectra of the specified DWD population
+#                    ## need to use the same frequencies as during the data generation process
+#                    N_spec=int(self.params['fs']*self.params['dur'])
+#                    fs_spec = np.fft.rfftfreq(N_spec, 1.0/self.params['fs'])[1:]
+#                    Sgw_fine = self.pop2spec(self.inj['popfile'],fs_spec,self.params['dur']*u.s,plot=False,names=self.inj['columns'],sep=self.inj['delimiter'])*4 ##h^2 = 1/2S_A = 1/2 * 1/2S_GW
+#                    ## now downsample to the frequencies at which we've evaluated the response
+#                    interp = interp1d(fs_spec,Sgw_fine)
+#                    Sgw = interp(self.fdata)
+#                else:
+#                    ## power spectra for analytic cases
+#                    if self.inj['spectral_inj'] == 'powerlaw':
+#                        Omega0, alpha = 10**self.inj['log_omega0'], self.inj['alpha']
+#                        Omegaf = Omega0*(self.fdata/self.params['fref'])**alpha
+#                    elif self.inj['spectral_inj'] == 'broken_powerlaw':
+#                        alpha_2 = self.inj['alpha1'] - 0.667
+#                        Omegaf = ((10**self.inj['log_A1'])*(self.fdata/self.params['fref'])**self.inj['alpha1'])/(\
+#                             1 + (10**self.inj['log_A2'])*(self.fdata/self.params['fref'])**alpha_2)
+#                    elif self.inj['spectral_inj'] == 'free_broken_powerlaw':
+#                        Omegaf = ((10**self.inj['log_A1'])*(self.fdata/self.params['fref'])**self.inj['alpha1'])/(\
+#                             1 + (10**self.inj['log_A2'])*(self.fdata/self.params['fref'])**self.inj['alpha2'])
+#                    elif self.inj['spectral_inj'] == 'broken_powerlaw_2':
+#                        delta = 0.1
+#                        Omegaf = (10**self.inj['log_omega0'])*(self.fdata/self.inj['f_break'])**(self.inj['alpha1']) \
+#                                * (0.5*(1+(self.fdata/self.inj['f_break'])**(1/delta)))**((self.inj['alpha1']-self.inj['alpha2'])*delta)
+#                    elif self.inj['spectral_inj'] == 'truncated_broken_powerlaw':
+#                        delta = 0.1
+#                        Omegaf = (10**self.inj['log_omega0'])*(self.fdata/self.inj['f_break'])**(self.inj['alpha1']) \
+#                                * (0.5*(1+(self.fdata/self.inj['f_break'])**(1/delta)))**((self.inj['alpha1']-self.inj['alpha2'])*delta) \
+#                                * 0.5 * (1+np.tanh((self.inj['f_cut']-self.fdata)/self.inj['f_scale']))
+#                    elif self.inj['spectral_inj'] == 'truncated_powerlaw':
+#                        Omegaf = (10**self.inj['log_omega0']) * (self.fdata/self.params['fref'])**self.inj['alpha'] \
+#                                * 0.5 * (1+np.tanh((self.inj['f_cut']-self.fdata)/self.inj['f_scale']))
+#    
+#                    Sgw = (3.0*(H0**2)*Omegaf)/(4*np.pi*np.pi*self.fdata**3)            
+            
+#            plt.close()
+#            if self.inj['injtype'] == 'multi':
+#                # Spectrum of each SGWB signal convoluted with their respective detector response tensor.
+#                S1_gw_a, S2_gw_a, S3_gw_a = Sgw_a[:, None]*R1_a, Sgw_a[:, None]*R2_a, Sgw_a[:, None]*R3_a
+#                S1_gw_i, S2_gw_i, S3_gw_i = Sgw_i[:, None]*R1_i, Sgw_i[:, None]*R2_i, Sgw_i[:, None]*R3_i
+#                
+#                # The total noise spectra is the sum of the instrumental + astrophysical (anisotropic and isotropic alike)
+#                S1, S2, S3 = S1[:, None] + S1_gw_a + S1_gw_i, S2[:, None] + S2_gw_a + S2_gw_i, S3[:, None] + S3_gw_a + S3_gw_i
+#                
+#                plt.loglog(self.fdata, np.mean(S1_gw_a,axis=1), label='Simulated anisotropic GW spectrum', lw=0.75)
+#                plt.loglog(self.fdata, np.mean(S1_gw_i,axis=1), label='Simulated isotropic GW spectrum', lw=0.75)
+#            else:
+#                # Spectrum of the SGWB signal convoluted with the detector response tensor.
+#                S1_gw, S2_gw, S3_gw = Sgw[:, None]*R1, Sgw[:, None]*R2, Sgw[:, None]*R3
+#
+#                # The total noise spectra is the sum of the instrumental + astrophysical
+#                S1, S2, S3 = S1[:, None] + S1_gw, S2[:, None] + S2_gw, S3[:, None] + S3_gw
+#                
+#                plt.loglog(self.fdata, np.mean(S1_gw,axis=1), label='Simulated GW spectrum', lw=0.75)
 
-        if self.inj['injtype'] != 'noise_only':
-            if self.inj['injtype'] == 'sph_sgwb' or self.inj['injtype']=='astro':
-                ## Response matrix : shape (3 x 3 x freq x time) if isotropic
-                ## set up different use cases
-                if self.inj['injtype'] == 'astro':
-                    if self.inj['injbasis'] == 'sph':
-                        summ_response_mat = np.sum(self.add_astro_signal(self.f0,self.tsegmid)*self.alms_inj[None,None,None,None,:],axis=-1)
-                    elif self.inj['injbasis'] == 'sph_lmax':
-                        summ_response_mat = np.sum(self.add_astro_signal(self.f0,self.tsegmid,set_almax=self.inj_almax)*self.alms_inj[None,None,None,None,:],axis=-1)
-                    elif self.inj['injbasis'] == 'pixel':
-                        summ_response_mat = self.add_astro_signal(self.f0,self.tsegmid,self.skymap_inj)  
-                else:
-                    summ_response_mat = np.sum(self.response_mat*self.alms_inj[None, None, None, None, :], axis=-1)
-                
-                # extra auto-power GW responses
-                R1 = np.real(summ_response_mat[0, 0, :, :])
-                R2 = np.real(summ_response_mat[1, 1, :, :])
-                R3 = np.real(summ_response_mat[2, 2, :, :])
-            
-            elif self.inj['injtype'] == 'multi':
-                if self.inj['injbasis'] == 'sph':
-                    summ_response_mat_a = np.sum(self.add_astro_signal_a(self.f0,self.tsegmid)*self.alms_inj[None,None,None,None,:],axis=-1)
-                    response_mat_i = self.add_astro_signal_i(self.f0, self.tsegmid)
-                    R1_a = np.real(summ_response_mat_a[0, 0, :, :])
-                    R2_a = np.real(summ_response_mat_a[1, 1, :, :])
-                    R3_a = np.real(summ_response_mat_a[2, 2, :, :])
-                    R1_i = np.real(response_mat_i[0, 0, :, :])
-                    R2_i = np.real(response_mat_i[1, 1, :, :])
-                    R3_i = np.real(response_mat_i[2, 2, :, :])
-                else:
-                    raise ValueError("Multi-SGWB injections only supports sph basis injections for now.")
-            else:
-                # extra auto-power GW responses
-                R1 = np.real(self.add_astro_signal(self.f0, self.tsegmid)[0, 0, :, :])
-                R2 = np.real(self.add_astro_signal(self.f0, self.tsegmid)[1, 1, :, :])
-                R3 = np.real(self.add_astro_signal(self.f0, self.tsegmid)[2, 2, :, :])
             
             
-            # Hubble constant
-            H0 = 2.2*10**(-18)
-            if self.inj['injtype'] == 'multi':
-                Omegaf_i = (10**self.inj['log_omega0_i'])*(self.fdata/self.params['fref'])**self.inj['alpha_i']
-                Omegaf_a = (10**self.inj['log_omega0_a']) * (self.fdata/self.params['fref'])**self.inj['alpha_a'] \
-                                * 0.5 * (1+np.tanh((self.inj['f_cut_a']-self.fdata)/self.inj['f_scale_a']))
-                Sgw_a = (3.0*(H0**2)*Omegaf_a)/(4*np.pi*np.pi*self.fdata**3)
-                Sgw_i = (3.0*(H0**2)*Omegaf_i)/(4*np.pi*np.pi*self.fdata**3)
-            else:
-                # Calculate astrophysical power law noise
-                if self.inj['spectral_inj'] == 'population':
-                    ## population-derived power spectra
-                    Sgw = self.pop2spec(self.inj['popfile'],self.fdata,self.params['dur']*u.s,names=self.inj['columns'],sep=self.inj['delimiter'])*4 ##h^2 = 1/2S_A = 1/2 * 1/2S_GW
-                    # Power spectra of the specified DWD population
-                    ## need to use the same frequencies as during the data generation process
-                    N_spec=int(self.params['fs']*self.params['dur'])
-                    fs_spec = np.fft.rfftfreq(N_spec, 1.0/self.params['fs'])[1:]
-                    Sgw_fine = self.pop2spec(self.inj['popfile'],fs_spec,self.params['dur']*u.s,plot=False,names=self.inj['columns'],sep=self.inj['delimiter'])*4 ##h^2 = 1/2S_A = 1/2 * 1/2S_GW
-                    ## now downsample to the frequencies at which we've evaluated the response
-                    interp = interp1d(fs_spec,Sgw_fine)
-                    Sgw = interp(self.fdata)
-                else:
-                    ## power spectra for analytic cases
-                    if self.inj['spectral_inj'] == 'powerlaw':
-                        Omega0, alpha = 10**self.inj['log_omega0'], self.inj['alpha']
-                        Omegaf = Omega0*(self.fdata/self.params['fref'])**alpha
-                    elif self.inj['spectral_inj'] == 'broken_powerlaw':
-                        alpha_2 = self.inj['alpha1'] - 0.667
-                        Omegaf = ((10**self.inj['log_A1'])*(self.fdata/self.params['fref'])**self.inj['alpha1'])/(\
-                             1 + (10**self.inj['log_A2'])*(self.fdata/self.params['fref'])**alpha_2)
-                    elif self.inj['spectral_inj'] == 'free_broken_powerlaw':
-                        Omegaf = ((10**self.inj['log_A1'])*(self.fdata/self.params['fref'])**self.inj['alpha1'])/(\
-                             1 + (10**self.inj['log_A2'])*(self.fdata/self.params['fref'])**self.inj['alpha2'])
-                    elif self.inj['spectral_inj'] == 'broken_powerlaw_2':
-                        delta = 0.1
-                        Omegaf = (10**self.inj['log_omega0'])*(self.fdata/self.inj['f_break'])**(self.inj['alpha1']) \
-                                * (0.5*(1+(self.fdata/self.inj['f_break'])**(1/delta)))**((self.inj['alpha1']-self.inj['alpha2'])*delta)
-                    elif self.inj['spectral_inj'] == 'truncated_broken_powerlaw':
-                        delta = 0.1
-                        Omegaf = (10**self.inj['log_omega0'])*(self.fdata/self.inj['f_break'])**(self.inj['alpha1']) \
-                                * (0.5*(1+(self.fdata/self.inj['f_break'])**(1/delta)))**((self.inj['alpha1']-self.inj['alpha2'])*delta) \
-                                * 0.5 * (1+np.tanh((self.inj['f_cut']-self.fdata)/self.inj['f_scale']))
-                    elif self.inj['spectral_inj'] == 'truncated_powerlaw':
-                        Omegaf = (10**self.inj['log_omega0']) * (self.fdata/self.params['fref'])**self.inj['alpha'] \
-                                * 0.5 * (1+np.tanh((self.inj['f_cut']-self.fdata)/self.inj['f_scale']))
-    
-                    Sgw = (3.0*(H0**2)*Omegaf)/(4*np.pi*np.pi*self.fdata**3)            
-            
-            plt.close()
-            if self.inj['injtype'] == 'multi':
-                # Spectrum of each SGWB signal convoluted with their respective detector response tensor.
-                S1_gw_a, S2_gw_a, S3_gw_a = Sgw_a[:, None]*R1_a, Sgw_a[:, None]*R2_a, Sgw_a[:, None]*R3_a
-                S1_gw_i, S2_gw_i, S3_gw_i = Sgw_i[:, None]*R1_i, Sgw_i[:, None]*R2_i, Sgw_i[:, None]*R3_i
-                
-                # The total noise spectra is the sum of the instrumental + astrophysical (anisotropic and isotropic alike)
-                S1, S2, S3 = S1[:, None] + S1_gw_a + S1_gw_i, S2[:, None] + S2_gw_a + S2_gw_i, S3[:, None] + S3_gw_a + S3_gw_i
-                
-                plt.loglog(self.fdata, np.mean(S1_gw_a,axis=1), label='Simulated anisotropic GW spectrum', lw=0.75)
-                plt.loglog(self.fdata, np.mean(S1_gw_i,axis=1), label='Simulated isotropic GW spectrum', lw=0.75)
-            else:
-                # Spectrum of the SGWB signal convoluted with the detector response tensor.
-                S1_gw, S2_gw, S3_gw = Sgw[:, None]*R1, Sgw[:, None]*R2, Sgw[:, None]*R3
-
-                # The total noise spectra is the sum of the instrumental + astrophysical
-                S1, S2, S3 = S1[:, None] + S1_gw, S2[:, None] + S2_gw, S3[:, None] + S3_gw
-                
-                plt.loglog(self.fdata, np.mean(S1_gw,axis=1), label='Simulated GW spectrum', lw=0.75)
-
-            
-            
-            plt.loglog(self.fdata, np.mean(S1,axis=1), label='Simulated Total spectrum', lw=0.75)
+            plt.loglog(self.fdata, S1, label='Simulated Total spectrum', lw=0.75)
 
 
         # noise budget plot
         plt.loglog(psdfreqs, data_PSD3,label='PSD, data series', alpha=0.6, lw=0.75)
         plt.loglog(self.fdata, C_noise[2, 2, :], label='Simulated instrumental noise spectrum', lw=0.75 )
         ## multi-SGWB injection plot gets squished due to truncated pl
-        if self.inj['injtype'] == 'multi':
-            ymin = 0.5 * S1_gw_i.min()
-            plt.ylim(bottom=ymin)
+#        if self.inj['injtype'] == 'multi':
+#            ymin = 0.5 * S1_gw_i.min()
+#            plt.ylim(bottom=ymin)
         ## population injection drops to zero inside frequency range and squishes the plot
-        elif self.inj['spectral_inj'] == 'population':
-            ymin = 0.5*S1_gw[S1_gw>1e-50].min()
-            plt.ylim(bottom=ymin)
+#        elif self.inj['spectral_inj'] == 'population':
+#            ymin = 0.5*S1_gw[S1_gw>1e-50].min()
+#            plt.ylim(bottom=ymin)
         plt.legend()
         plt.xlabel('$f$ in Hz')
         plt.ylabel('PSD 1/Hz ')
@@ -404,7 +447,7 @@ class LISA(LISAdata, Model):
         plt.close()
 
 
-        plt.loglog(self.fdata, np.mean(S3,axis=1), label='required')
+        plt.loglog(self.fdata, S3, label='required')
         plt.loglog(psdfreqs, data_PSD3,label='PSD, data', alpha=0.6)
         plt.xlabel('$f$ in Hz')
         plt.ylabel('PSD 1/Hz ')
@@ -420,7 +463,7 @@ class LISA(LISAdata, Model):
 
 
         ## lets also plot psd residue.
-        rel_res_mean = (data_PSD3 - np.mean(S3,axis=1))/np.mean(S3,axis=1)
+        rel_res_mean = (data_PSD3 - S3)/S3
 
         plt.semilogx(self.fdata, rel_res_mean , label='relative mean residue')
         plt.xlabel('f in Hz')
@@ -437,47 +480,49 @@ class LISA(LISAdata, Model):
         # cross-power diag plots. We will only do 12. IF TDI=XYZ this is S_XY and if TDI=AET
         # this will be S_AE
 
-        ii, jj = 2, 0
-
-        if self.inj['injtype'] == 'noise_only':
-            Sx = C_noise[ii, jj, :]
-        elif self.inj['injtype'] == 'multi':
-            Sx = C_noise[ii, jj, :, None] + Sgw_i[:,None]*response_mat_i[ii,jj,:,0] + Sgw_a*summ_response_mat_a[ii, jj, :, 0]
-        elif self.inj['injtype'] == 'sph_sgwb' or self.inj['injtype'] == 'astro':
-            Sx = C_noise[ii, jj, :] + Sgw*summ_response_mat[ii, jj, :, 0]
-        else:
-            Sx = C_noise[ii, jj, :, None] + Sgw[:,None]*self.response_mat[ii, jj, :, 0]
-
-        CSDx = np.mean(np.conj(self.rbar[:, :, ii]) * self.rbar[:, :, jj], axis=1)
-
-        plt.subplot(2, 1, 1)
-        if len(Sx.shape) == 1:
-            plt.loglog(self.fdata, np.abs(np.real(Sx)), label='Re(Required ' + str(ii+1) + str(jj+1) + ')' )
-        else:
-            plt.loglog(self.fdata, np.mean(np.abs(np.real(Sx)),axis=1), label='Re(Required ' + str(ii+1) + str(jj+1) + ')' )
-        plt.loglog(psdfreqs, np.abs(np.real(CSDx)) ,label='Re(CSD' + str(ii+1) + str(jj+1) + ')', alpha=0.6)
-        plt.xlabel('f in Hz')
-        plt.ylabel('Power in 1/Hz')
-        plt.legend()
-        plt.ylim([1e-44, 5e-40])
-        plt.xlim(0.5*self.params['fmin'], 2*self.params['fmax'])
-        plt.grid()
-
-        plt.subplot(2, 1, 2)
-        if len(Sx.shape) == 1:
-            plt.loglog(self.fdata, np.abs(np.imag(Sx)), label='Im(Required ' + str(ii+1) + str(jj+1) + ')' )
-        else:
-            plt.loglog(self.fdata, np.mean(np.abs(np.imag(Sx)),axis=1), label='Im(Required ' + str(ii+1) + str(jj+1) + ')' )
-        plt.loglog(psdfreqs, np.abs(np.imag(CSDx)) ,label='Im(CSD' + str(ii+1) + str(jj+1) + ')', alpha=0.6)
-        plt.xlabel('f in Hz')
-        plt.ylabel(' Power in 1/Hz')
-        plt.legend()
-        plt.xlim(0.5*self.params['fmin'], 2*self.params['fmax'])
-        plt.ylim([1e-44, 5e-40])
-        plt.grid()
-        plt.savefig(self.params['out_dir'] + '/diag_csd_' + str(ii+1) + str(jj+1) + '.png', dpi=200)
-        print('Diagnostic spectra plot made in ' + self.params['out_dir'] + '/diag_csd_' + str(ii+1) + str(jj+1) + '.png')
-        plt.close()
+        ## FIX LATER
+        
+#        ii, jj = 2, 0
+#
+#        if self.inj['injtype'] == 'noise_only':
+#            Sx = C_noise[ii, jj, :]
+#        elif self.inj['injtype'] == 'multi':
+#            Sx = C_noise[ii, jj, :, None] + Sgw_i[:,None]*response_mat_i[ii,jj,:,0] + Sgw_a*summ_response_mat_a[ii, jj, :, 0]
+#        elif self.inj['injtype'] == 'sph_sgwb' or self.inj['injtype'] == 'astro':
+#            Sx = C_noise[ii, jj, :] + Sgw*summ_response_mat[ii, jj, :, 0]
+#        else:
+#            Sx = C_noise[ii, jj, :, None] + Sgw[:,None]*self.response_mat[ii, jj, :, 0]
+#
+#        CSDx = np.mean(np.conj(self.rbar[:, :, ii]) * self.rbar[:, :, jj], axis=1)
+#
+#        plt.subplot(2, 1, 1)
+#        if len(Sx.shape) == 1:
+#            plt.loglog(self.fdata, np.abs(np.real(Sx)), label='Re(Required ' + str(ii+1) + str(jj+1) + ')' )
+#        else:
+#            plt.loglog(self.fdata, np.mean(np.abs(np.real(Sx)),axis=1), label='Re(Required ' + str(ii+1) + str(jj+1) + ')' )
+#        plt.loglog(psdfreqs, np.abs(np.real(CSDx)) ,label='Re(CSD' + str(ii+1) + str(jj+1) + ')', alpha=0.6)
+#        plt.xlabel('f in Hz')
+#        plt.ylabel('Power in 1/Hz')
+#        plt.legend()
+#        plt.ylim([1e-44, 5e-40])
+#        plt.xlim(0.5*self.params['fmin'], 2*self.params['fmax'])
+#        plt.grid()
+#
+#        plt.subplot(2, 1, 2)
+#        if len(Sx.shape) == 1:
+#            plt.loglog(self.fdata, np.abs(np.imag(Sx)), label='Im(Required ' + str(ii+1) + str(jj+1) + ')' )
+#        else:
+#            plt.loglog(self.fdata, np.mean(np.abs(np.imag(Sx)),axis=1), label='Im(Required ' + str(ii+1) + str(jj+1) + ')' )
+#        plt.loglog(psdfreqs, np.abs(np.imag(CSDx)) ,label='Im(CSD' + str(ii+1) + str(jj+1) + ')', alpha=0.6)
+#        plt.xlabel('f in Hz')
+#        plt.ylabel(' Power in 1/Hz')
+#        plt.legend()
+#        plt.xlim(0.5*self.params['fmin'], 2*self.params['fmax'])
+#        plt.ylim([1e-44, 5e-40])
+#        plt.grid()
+#        plt.savefig(self.params['out_dir'] + '/diag_csd_' + str(ii+1) + str(jj+1) + '.png', dpi=200)
+#        print('Diagnostic spectra plot made in ' + self.params['out_dir'] + '/diag_csd_' + str(ii+1) + str(jj+1) + '.png')
+#        plt.close()
         
     def plot_spectra(self):
         '''
@@ -555,10 +600,18 @@ def blip(paramsfile='params.ini',resume=False):
 
     # Injection Dict
     inj['doInj']       = int(config.get("inj", "doInj"))
+    
+    inj['injection']       = str(config.get("inj", "injection"))
+    
     inj['injtype']     = str(config.get("inj", "injtype"))
     
     inj['log_Np']      = np.log10(float(config.get("inj", "Np")))
     inj['log_Na']      = np.log10(float(config.get("inj", "Na")))
+    
+    
+#    inj.extend(parameter_collector(inj['injection']))
+    ## need to write parameter_collector (in models.py) that automatically looks for the relevant parameters in the params file given the model
+    
     
     ## spectral parameters
     if inj['injtype'] != 'noise_only':
