+++ src/makeLISAdata.py
@@ -4,7 +4,6 @@ from src.instrNoise import instrNoise
 from src.geometry import geometry
 from src.sph_geometry import sph_geometry
 from src.populations import populations
-from scipy.interpolate import interp1d as intrp
 import matplotlib.pyplot as plt
 import healpy as hp
 from astropy import units as u
@@ -55,479 +54,136 @@ class LISAdata(geometry, sph_geometry, instrNoise, populations):
             return r1, r2, r3, fdata
 
 
-    def gaussianData(self, Sh,freqs, fs=1, dur=1e5):
-
-        '''
-        Script for generation time series noise drawn from a gaussian process of a given spectral density.  Adapted from gaussian_noise.m from stamp
-
-        Parameters
-        -----------
-
-        Sh : (float)
-            A frequency array with the desired power spectral density
-        freqs : (float)
-            An array with corresponding frequencies to Sh
-
-        fs : (float)
-            SampleRate in Hz
-
-        dur : (int)
-            Duration in seconds
-
-
-        Returns
-        ---------
-
-        ht : float
-        Array with time series data of duration, dur with the prescribed spectrum Sh
-
-
-        '''
-
-        # Number of data points in the time series
-        N = int(fs*dur)
-
-        # prepare for FFT
-        if  np.mod(N,2)== 0 :
-            numFreqs = int(N/2 - 1)
-        else:
-            numFreqs = int((N-1)/2)
-
-        # We will make an array of the desired frequencies
-        delF = 1/dur
-        fmin = 0
-        fmax = np.around(dur*fs/2)/dur
-        delF = 1/dur
-
-        # The output frequency series
-        fout = np.linspace(fmin, fmax, numFreqs)
-
-        # Interpolate to the desired frequencies
-        norms = np.interp(fout, freqs, Sh)
-
-        # Amplitude for for ifft
-        norms = np.sqrt(norms*fs*N)/2.0
-
-        # Normally distributed in frequency space
-        re1 = norms*np.random.normal(size=fout.size)
-        im1 = norms*np.random.normal(size=fout.size)
-
-        htilda = re1 + 1j*im1
-
-        if np.mod(N, 2) == 0:
-            htilda = np.concatenate((np.zeros(1), htilda,np.zeros(1), np.flipud(np.conjugate(htilda))))
-        else:
-            htilda = np.concatenate((np.zeros(1),htilda, np.conjugate(np.flipud(htilda))))
-
-        # Take inverse fft to get time series data
-        ht = np.real(np.fft.ifft(htilda, N))
-
-        return ht
-
-    def freqdomain_gaussianData(self, Sh,freqs, fs=1, dur=1e5):
-
-        '''
-        Script to generate freq Domain gaussian data of a given spectral density.
-
-        Parameters
-        -----------
-
-        Sh : (float)
-            A frequency array with the desired power spectral density
-        freqs : (float)
-            An array with corresponding frequencies to Sh
-
-        fs : (float)
-            SampleRate in Hz
-
-        dur : (int)
-            Duration in seconds
-
-
-        Returns
-        ---------
-
-        ht : float
-        frequency domain gaussian.
-        '''
-
-        # Number of data points in the time series
-        N = int(fs*dur)
-
-        # prepare for FFT
-        if  np.mod(N,2)== 0 :
-            numFreqs = N/2 - 1;
-        else:
-            numFreqs = (N-1)/2;
-
-        # We will make an array of the desired frequencies
-        delF = 1/dur
-        fmin = 1/dur
-        fmax = np.around(dur*fs/2)/dur
-        delF = 1/dur
-
-        # The output frequency series
-        fout = np.linspace(fmin, fmax, numFreqs)
-
-        # Interpolate to the desired frequencies
-        norms = np.interp(fout, freqs, Sh)
-
-        # Amplitude for for ifft
-        norms = np.sqrt(norms*fs*N)/2.0
-
-        # Normally distributed in frequency space
-        re1 = norms*np.random.normal(size=fout.size)
-        im1 = norms*np.random.normal(size=fout.size)
-
-        htilda = re1 + 1j*im1
-
-
-        return htilda, fout
-
-
-    def gen_michelson_noise(self):
-
-        '''
-        Generate interferometric michelson (time-domain) noise, using freqDomain.fundamental_noise_spectrum
-
-        Returns
-        ---------
-        h1, h2, h3 : float
-            Time series data for the three michelson channels
-        '''
-
-        # --------------------- Generate Fake Noise -----------------------------
-        print("Simulating instrumental noise ...")
-
-       # speed of light
-        cspeed = 3e8 #m/s
-        delf  = 1.0/self.params['dur']
-        frange = np.arange(self.params['fmin'], self.params['fmax'], delf) # in Hz
-        fstar = 3e8/(2*np.pi*self.armlength)
-        f0 = frange/(2*fstar)
-
-        Sp, Sa = self.fundamental_noise_spectrum(frange, Np=10**self.inj['log_Np'], Na=10**self.inj['log_Na'])
-
-        # Generate data
-        np12 = self.gaussianData(Sp, frange, self.params['fs'], 1.1*self.params['dur'])
-        np21 = self.gaussianData(Sp, frange, self.params['fs'], 1.1*self.params['dur'])
-        np13 = self.gaussianData(Sp, frange, self.params['fs'], 1.1*self.params['dur'])
-        np31 = self.gaussianData(Sp, frange, self.params['fs'], 1.1*self.params['dur'])
-        np23 = self.gaussianData(Sp, frange, self.params['fs'], 1.1*self.params['dur'])
-        np32 = self.gaussianData(Sp, frange, self.params['fs'], 1.1*self.params['dur'])
-
-        na12 = self.gaussianData(Sa, frange, self.params['fs'], 1.1*self.params['dur'])
-        na21 = self.gaussianData(Sa, frange, self.params['fs'], 1.1*self.params['dur'])
-        na13 = self.gaussianData(Sa, frange, self.params['fs'], 1.1*self.params['dur'])
-        na31 = self.gaussianData(Sa, frange, self.params['fs'], 1.1*self.params['dur'])
-        na23 = self.gaussianData(Sa, frange, self.params['fs'], 1.1*self.params['dur'])
-        na32 = self.gaussianData(Sa, frange, self.params['fs'], 1.1*self.params['dur'])
-
-        # time array and time shift array
-        tarr =  np.arange(0, 1.1*self.params['dur'], 1.0/self.params['fs'])
-        tarr = tarr[0:np12.size]
-        delt = tarr[2] - tarr[1]
-
-        # We start with assuming a padding of 20 seconds on the beginning for the
-        # Michelson channels
-        ## Using up ten seconds here.
-        ten_idx = int(self.params['fs']*10)
-
-        # To implement TDI we need time shifts of multiples of L.
-        tlag  = self.armlength/cspeed
-
-        ## One way dopper channels for each arms. Using up seconds of the pad here for doing tlag
-        f21 = intrp(tarr, na21, kind='cubic', fill_value='extrapolate')
-        f12 = intrp(tarr, na12, kind='cubic', fill_value='extrapolate')
-        f32 = intrp(tarr, na32, kind='cubic', fill_value='extrapolate')
-        f23 = intrp(tarr, na23, kind='cubic', fill_value='extrapolate')
-        f13 = intrp(tarr, na13, kind='cubic', fill_value='extrapolate')
-        f31 = intrp(tarr, na31, kind='cubic', fill_value='extrapolate')
-
-        h12  = np12[ten_idx:] - na12[ten_idx:] + f21(tarr[ten_idx:]-tlag)
-        h21  = np21[ten_idx:] + na21[ten_idx:] - f12(tarr[ten_idx:]-tlag)
-
-        h23  = np23[ten_idx:] - na23[ten_idx:] + f32(tarr[ten_idx:]-tlag)
-        h32  = np32[ten_idx:] + na32[ten_idx:] - f23(tarr[ten_idx:]-tlag)
-
-        h31  = np31[ten_idx:] - na31[ten_idx:] + f13(tarr[ten_idx:]-tlag)
-        h13  = np13[ten_idx:] + na13[ten_idx:] - f31(tarr[ten_idx:]-tlag)
-
-        ## reduce tarr
-        tarr = tarr[ten_idx:]
-
-        # The Michelson channels, formed from the doppler channels. Using the other
-        # ten seconds here
-
-        f12 = intrp(tarr, h12, kind='cubic', fill_value='extrapolate')
-        f13 = intrp(tarr, h13, kind='cubic', fill_value='extrapolate')
-        f23 = intrp(tarr, h23, kind='cubic', fill_value='extrapolate')
-        f21 = intrp(tarr, h21, kind='cubic', fill_value='extrapolate')
-        f31 = intrp(tarr, h31, kind='cubic', fill_value='extrapolate')
-        f32 = intrp(tarr, h32, kind='cubic', fill_value='extrapolate')
-
-
-        h1 = f12(tarr[ten_idx:]-tlag) + h21[ten_idx:] - \
-                f13(tarr[ten_idx:]-tlag)  - h31[ten_idx:]
-
-        h2 = f23(tarr[ten_idx:]-tlag) + h32[ten_idx:] - \
-                f21(tarr[ten_idx:]-tlag)  - h12[ten_idx:]
-
-        h3 = f31(tarr[ten_idx:]-tlag)  + h13[ten_idx:] - \
-                f32(tarr[ten_idx:]-tlag)  - h23[ten_idx:]
-
-
-        '''
-        Older way of doing time shifts is commented out here. Interp doesn't work since it
-        creates correlated samples, but I leave it here for reference. - Sharan
-
-        h1 = np.interp(tshift, tarr, h12, left=h12[0]) + h21 -\
-        np.interp(tshift, tarr, h13, left=h13[0]) - h31
-
-        h2 = np.interp(tshift, tarr, h23, left=h23[0]) + h32 -\
-        np.interp(tshift, tarr, h21, left=h21[0]) - h12
-
-        h3 = np.interp(tshift, tarr, h31, left=h31[0]) + h13 -\
-        np.interp(tshift, tarr, h32, left=h32[0]) - h23
-        '''
-
-        return tarr[ten_idx:], h1, h2, h3
-
-
-
-    def gen_xyz_noise(self):
-
-        '''
-        Generate interferometric A, E and T channel TDI (time-domain) noise, using freqDomain.fundamental_noise_spectrum
-
-        Returns
-        ---------
-
-        h1_noi, h2_noi, h3_noi : float
-            Time series data for the three TDI channels
-
-        '''
-
-        '''
-
-        '''
-        cspeed = 3e8 #m/s
-
-        # michelson channels
-        tarr, hm1, hm2, hm3 = self.gen_michelson_noise()
-
-        ## Using up ten seconds here.
-        ten_idx = int(self.params['fs']*10)
-
-        # Introduce time series
-        tshift = 2*self.armlength/cspeed
-
-        f1 = intrp(tarr, hm1, kind='cubic', fill_value='extrapolate')
-        f2 = intrp(tarr, hm2, kind='cubic', fill_value='extrapolate')
-        f3 = intrp(tarr, hm3, kind='cubic', fill_value='extrapolate')
-
-
-        hX = hm1[ten_idx:] - f1(tarr[ten_idx:] - tshift)
-        hY = hm2[ten_idx:] - f2(tarr[ten_idx:] - tshift)
-        hZ = hm3[ten_idx:] - f3(tarr[ten_idx:] - tshift)
-
-        return tarr[ten_idx:], hX, hY, hZ
-
-
-
-    def gen_aet_noise(self):
-
-        '''
-        Generate interferometric A, E and T channel TDI (time-domain) noise, using freqDomain.fundamental_noise_spectrum
-
-        Returns
-        ---------
-        h1_noi, h2_noi, h3_noi : float
-            Time series data for the three TDI channels
-
-        '''
-
-        cspeed = 3e8 #m/s
-
-        # michelson channels
-        tarr, hX, hY, hZ = self.gen_xyz_noise()
-
-        h1_noi = (1.0/3.0)*(2*hX - hY - hZ)
-        h2_noi = (1.0/np.sqrt(3.0))*(hZ - hY)
-        h3_noi = (1.0/3.0)*(hX + hY + hZ)
-
-        return tarr, h1_noi, h2_noi, h3_noi
-
-
-    def gen_noise_cov(self):
-
-        '''
-        Generate interferometric (time-domain) noise, using a frequency domain covariance
-        spectrum matrix rather than time delays in time domain.
-        ---------
-
-        h1_noi, h2_noi, h3_noi : float
-            Time series data for the three TDI channels
-        '''
-
-        cspeed = 3e8 #m/s
-        delf  = 1.0/self.params['dur']
-        fstar = 3e8/(2*np.pi*self.armlength)
-        N = int(self.params['fs']*self.params['dur'])
-
-        frange = np.fft.rfftfreq(N, 1.0/self.params['fs'])[1:]
-        frange = frange[frange <= self.params['fmax']]
-        frange = frange[frange >= self.params['fmin']]
-
-        f0 = frange/(2*fstar)
-
-        #Sp, Sa = self.fundamental_noise_spectrum(frange, Np=10**self.inj['log_Np'], Na=10**self.inj['log_Na'])
-
-        C_xyz = self.xyz_noise_spectrum(frange, f0, Np=10**self.inj['log_Np'], Na=10**self.inj['log_Na'])
-
-        ## Cholesky decomposition to get the "sigma" matrix
-        L_cholesky = np.sqrt(self.params['fs'] * N/4.0) *  np.linalg.cholesky(np.moveaxis(C_xyz, -1, 0))
-
-        #for ii in range(C_xyz.shape[-1]):
-        #    np.linalg.cholesky(C_xyz[:, :, ii])
-        #    print(str(ii) + '/' + str(C_xyz.shape[-1]))
-
-        ## generate standard normal complex data frist
-        z_norm = np.random.normal(size=(3, frange.size)) + 1j * np.random.normal(size=(3, frange.size))
-
-        ## initialize a new scaled array. The data in z_norm will be rescaled into z_scale
-        z_scale = np.zeros(z_norm.shape, dtype='complex')
-
-        for ii in range(frange.size):
-            z_scale[:, ii] = np.matmul(L_cholesky[ii, :, :], z_norm[:, ii])
-
-
-        ## The three channels : concatenate with norm at f = 0 to be zero
-        htilda1  = np.concatenate([ [0], z_scale[0, :]])
-        htilda2  = np.concatenate([ [0], z_scale[1, :]])
-        htilda3  = np.concatenate([ [0], z_scale[2, :]])
-
-
-        # Take inverse fft to get time series data
-        h1 = np.fft.irfft(htilda1, N)
-        h2 = np.fft.irfft(htilda2, N)
-        h3 = np.fft.irfft(htilda3, N)
-
-        tarr =  np.arange(0, self.params['dur'], 1.0/self.params['fs'])
-
-
-        return tarr, h1, h2, h3
-
-    def add_sgwb_data(self, tbreak = 0.0,multi='no'):
-
-        cspeed = 3e8 #m/s
-
-        ## define the splice segment duration
-        tsplice = 1e4
-        delf  = 1.0/tsplice
-
-        ## the segments to be splices are half-overlapping
-        nsplice = 2*int(self.params['dur']/tsplice) + 1
-
-        ## arrays of segmnent start and mid times
-        #tmids = (tsplice/2.0) * np.arange(nsplice) + (tsplice/2.0)
-
-        ## arrays of segmnent start and mid times
-        tmids = self.params['tstart'] + tbreak +  (tsplice/2.0) * np.arange(nsplice) + (tsplice/2.0)
-
-        ## Number of time-domain points in a splice segment
-        N = int(self.params['fs']*tsplice)
-        halfN = int(0.5*N)
-
-        ## leave out f = 0
-        frange = np.fft.rfftfreq(N, 1.0/self.params['fs'])[1:]
-
-        ## the charecteristic frequency of LISA, and the scaled frequency array
-        fstar = 3e8/(2*np.pi*self.armlength)
-        f0 = frange/(2*fstar)
+    
+
+
+    
+
+    def add_sgwb_data(self, injmodel, tbreak = 0.0):
+        
+        
+        
+
+#        cspeed = 3e8 #m/s
+#
+#        ## define the splice segment duration
+#        tsplice = 1e4
+#        delf  = 1.0/tsplice
+#
+#        ## the segments to be splices are half-overlapping
+#        nsplice = 2*int(self.params['dur']/tsplice) + 1
+#
+#        ## arrays of segmnent start and mid times
+#        #tmids = (tsplice/2.0) * np.arange(nsplice) + (tsplice/2.0)
+#
+#        ## arrays of segmnent start and mid times
+#        tmids = self.params['tstart'] + tbreak +  (tsplice/2.0) * np.arange(nsplice) + (tsplice/2.0)
+#
+#        ## Number of time-domain points in a splice segment
+#        N = int(self.params['fs']*tsplice)
+#        halfN = int(0.5*N)
+#
+#        ## leave out f = 0
+#        frange = np.fft.rfftfreq(N, 1.0/self.params['fs'])[1:]
+#
+#        ## the charecteristic frequency of LISA, and the scaled frequency array
+#        fstar = 3e8/(2*np.pi*self.armlength)
+#        f0 = frange/(2*fstar)
 
         ## Response matrix : shape (3 x 3 x freq x time) if isotropic
         ## set up different use cases
         ## pixel basis computes response in conjunction with skymap, so skip this step
-        if self.inj['injtype'] == 'multi':
-            if multi == 'a':
-                response_mat = self.add_astro_signal_a(f0,tmids)
-            elif multi == 'i':
-                response_mat = self.add_astro_signal_i(f0,tmids)
-            else:
-                raise ValueError("Invalid specification of 'multi' type. Can be 'a' (anisotropic) or 'i' (isotropic).")
-        elif self.inj['injtype'] == 'astro' and self.inj['injbasis'] == 'pixel':
-            pass
-        else:
-            if self.inj['injtype'] == 'astro' and self.inj['injbasis'] == 'sph_lmax':
-                signal_args= (f0,tmids,self.inj_almax)
-            else:
-                signal_args = (f0,tmids)
-        
-            response_mat = self.add_astro_signal(*signal_args)
+#        if self.inj['injtype'] == 'multi':
+#            if multi == 'a':
+#                response_mat = self.add_astro_signal_a(f0,tmids)
+#            elif multi == 'i':
+#                response_mat = self.add_astro_signal_i(f0,tmids)
+#            else:
+#                raise ValueError("Invalid specification of 'multi' type. Can be 'a' (anisotropic) or 'i' (isotropic).")
+#        elif self.inj['injtype'] == 'astro' and self.inj['injbasis'] == 'pixel':
+#            pass
+#        else:
+#            if self.inj['injtype'] == 'astro' and self.inj['injbasis'] == 'sph_lmax':
+#                signal_args= (f0,tmids,self.inj_almax)
+#            else:
+#                signal_args = (f0,tmids)
+#        
+#            response_mat = self.add_astro_signal(*signal_args)
 
         ## Cholesky decomposition to get the "sigma" matrix
-        H0 = 2.2*10**(-18) ## in SI units
+#        H0 = 2.2*10**(-18) ## in SI units
         
-        ## compute spectra and save Omega(1mHz) for later
-        if self.inj['injtype'] == 'multi':
-            if multi == 'a':
-                Omegaf = (10**self.inj['log_omega0_a']) * (frange/self.params['fref'])**self.inj['alpha_a'] \
-                                * 0.5 * (1+np.tanh((self.inj['f_cut_a']-frange)/self.inj['f_scale_a']))
-                Omega_1mHz = (10**self.inj['log_omega0_a']) * (1e-3/self.params['fref'])**self.inj['alpha_a'] \
-                                * 0.5 * (1+np.tanh((self.inj['f_cut_a']-1e-3)/self.inj['f_scale_a']))
-            elif multi == 'i':
-                Omegaf = (10**self.inj['log_omega0_i'])*(frange/self.params['fref'])**self.inj['alpha_i']
-                Omega_1mHz = (10**self.inj['log_omega0_i'])*(1e-3/self.params['fref'])**self.inj['alpha_i']
-            else:
-                raise ValueError("Invalid specification of multi. Can be 'a' (anisotropic) or 'i' (isotropic).")
-        elif self.inj['spectral_inj'] == 'powerlaw':
-            Omegaf = (10**self.inj['log_omega0'])*(frange/(self.params['fref']))**self.inj['alpha']
-            Omega_1mHz = 10**(self.inj['log_omega0']) * (1e-3/self.params['fref'])**(self.inj['alpha'])
-        elif self.inj['spectral_inj'] == 'broken_powerlaw':
-            alpha_2 = self.inj['alpha1'] - 0.667
-            Omegaf = ((10**self.inj['log_A1'])*(frange/self.params['fref'])**self.inj['alpha1'])/(\
-                     1 + (10**self.inj['log_A2'])*(frange/self.params['fref'])**alpha_2)
-            Omega_1mHz= ((10**self.inj['log_A1'])*(1e-3/self.params['fref'])**self.inj['alpha1'])/(\
-                        1 + (10**self.inj['log_A2'])*(1e-3/self.params['fref'])**alpha_2)
-        elif self.inj['spectral_inj'] == 'broken_powerlaw_2':
-            delta = 0.1
-            Omegaf = (10**self.inj['log_omega0'])*(frange/self.inj['f_break'])**(self.inj['alpha1']) \
-                    * (0.5*(1+(frange/self.inj['f_break'])**(1/delta)))**((self.inj['alpha1']-self.inj['alpha2'])*delta)
-            Omega_1mHz = (10**self.inj['log_omega0'])*(1e-3/self.inj['f_break'])**(self.inj['alpha1']) \
-                    * (0.5*(1+(1e-3/self.inj['f_break'])**(1/delta)))**((self.inj['alpha1']-self.inj['alpha2'])*delta)
-        elif self.inj['spectral_inj'] == 'free_broken_powerlaw':
-            Omegaf = ((10**self.inj['log_A1'])*(frange/self.params['fref'])**self.inj['alpha1'])/(\
-                     1 + (10**self.inj['log_A2'])*(frange/self.params['fref'])**self.inj['alpha2'])
-            Omega_1mHz = ((10**self.inj['log_A1'])*(1e-3/self.params['fref'])**self.inj['alpha1'])/(\
-                        1 + (10**self.inj['log_A2'])*(1e-3/self.params['fref'])**self.inj['alpha2'])
-        elif self.inj['spectral_inj'] == 'truncated_broken_powerlaw':
-            delta = 0.1
-            Omegaf = (10**self.inj['log_omega0'])*(frange/self.inj['f_break'])**(self.inj['alpha1']) \
-                    * (0.5*(1+(frange/self.inj['f_break'])**(1/delta)))**((self.inj['alpha1']-self.inj['alpha2'])*delta) \
-                    * 0.5 * (1 + np.tanh((self.inj['f_cut'] - frange)/self.inj['f_scale']))
-            Omega_1mHz = (10**self.inj['log_omega0'])*(1e-3/self.inj['f_break'])**(self.inj['alpha1']) \
-                    * (0.5*(1+(1e-3/self.inj['f_break'])**(1/delta)))**((self.inj['alpha1']-self.inj['alpha2'])*delta) \
-                    * 0.5 * (1 + np.tanh((self.inj['f_cut'] - 1e-3)/self.inj['f_scale']))
-        elif self.inj['spectral_inj'] == 'truncated_powerlaw':
-            Omegaf = (10**self.inj['log_omega0'])*(frange/(self.params['fref']))**self.inj['alpha'] \
-                    * 0.5 * (1 + np.tanh((self.inj['f_cut'] - frange)/self.inj['f_scale']))
-            Omega_1mHz = 10**(self.inj['log_omega0']) * (1e-3/self.params['fref'])**(self.inj['alpha']) \
-                    * 0.5 * (1 + np.tanh((self.inj['f_cut'] - 1e-3)/self.inj['f_scale']))
-
-        elif self.inj['spectral_inj'] == 'population':
-            print("Constructing foreground spectrum from DWD population...")
-            ## factor of two b/c (h_A,h_A*)~h^2~1/2 * S_A
-            ## additional factor of 2 b/c S_GW = 2 * S_A
-            Sgw = self.pop2spec(self.inj['popfile'],frange,self.params['dur']*u.s,names=self.inj['columns'],sep=self.inj['delimiter'])*4
-            Omega_1mHz = Sgw[np.argmin(np.abs(frange - 1e-3))]/((3/(4*(1e-3)**3))*(H0/np.pi)**2)
-        else:
-            raise ValueError("Unknown spectral injection selected. Can be powerlaw, broken_powerlaw, free_broken_powerlaw, or population.")
         
-        # Spectrum of the SGWB from Omegaf (population version goes directly to the spectrum from binary strains and frequencies)
-        if self.inj['spectral_inj'] != 'population':
-            Sgw = Omegaf*(3/(4*frange**3))*(H0/np.pi)**2    
-
+        
+        
+#        ## compute spectra and save Omega(1mHz) for later
+#        if self.inj['injtype'] == 'multi':
+#            if multi == 'a':
+#                Omegaf = (10**self.inj['log_omega0_a']) * (frange/self.params['fref'])**self.inj['alpha_a'] \
+#                                * 0.5 * (1+np.tanh((self.inj['f_cut_a']-frange)/self.inj['f_scale_a']))
+#                Omega_1mHz = (10**self.inj['log_omega0_a']) * (1e-3/self.params['fref'])**self.inj['alpha_a'] \
+#                                * 0.5 * (1+np.tanh((self.inj['f_cut_a']-1e-3)/self.inj['f_scale_a']))
+#            elif multi == 'i':
+#                Omegaf = (10**self.inj['log_omega0_i'])*(frange/self.params['fref'])**self.inj['alpha_i']
+#                Omega_1mHz = (10**self.inj['log_omega0_i'])*(1e-3/self.params['fref'])**self.inj['alpha_i']
+#            else:
+#                raise ValueError("Invalid specification of multi. Can be 'a' (anisotropic) or 'i' (isotropic).")
+#        elif self.inj['spectral_inj'] == 'powerlaw':
+#            Omegaf = (10**self.inj['log_omega0'])*(frange/(self.params['fref']))**self.inj['alpha']
+#            Omega_1mHz = 10**(self.inj['log_omega0']) * (1e-3/self.params['fref'])**(self.inj['alpha'])
+#        elif self.inj['spectral_inj'] == 'broken_powerlaw':
+#            alpha_2 = self.inj['alpha1'] - 0.667
+#            Omegaf = ((10**self.inj['log_A1'])*(frange/self.params['fref'])**self.inj['alpha1'])/(\
+#                     1 + (10**self.inj['log_A2'])*(frange/self.params['fref'])**alpha_2)
+#            Omega_1mHz= ((10**self.inj['log_A1'])*(1e-3/self.params['fref'])**self.inj['alpha1'])/(\
+#                        1 + (10**self.inj['log_A2'])*(1e-3/self.params['fref'])**alpha_2)
+#        elif self.inj['spectral_inj'] == 'broken_powerlaw_2':
+#            delta = 0.1
+#            Omegaf = (10**self.inj['log_omega0'])*(frange/self.inj['f_break'])**(self.inj['alpha1']) \
+#                    * (0.5*(1+(frange/self.inj['f_break'])**(1/delta)))**((self.inj['alpha1']-self.inj['alpha2'])*delta)
+#            Omega_1mHz = (10**self.inj['log_omega0'])*(1e-3/self.inj['f_break'])**(self.inj['alpha1']) \
+#                    * (0.5*(1+(1e-3/self.inj['f_break'])**(1/delta)))**((self.inj['alpha1']-self.inj['alpha2'])*delta)
+#        elif self.inj['spectral_inj'] == 'free_broken_powerlaw':
+#            Omegaf = ((10**self.inj['log_A1'])*(frange/self.params['fref'])**self.inj['alpha1'])/(\
+#                     1 + (10**self.inj['log_A2'])*(frange/self.params['fref'])**self.inj['alpha2'])
+#            Omega_1mHz = ((10**self.inj['log_A1'])*(1e-3/self.params['fref'])**self.inj['alpha1'])/(\
+#                        1 + (10**self.inj['log_A2'])*(1e-3/self.params['fref'])**self.inj['alpha2'])
+#        elif self.inj['spectral_inj'] == 'truncated_broken_powerlaw':
+#            delta = 0.1
+#            Omegaf = (10**self.inj['log_omega0'])*(frange/self.inj['f_break'])**(self.inj['alpha1']) \
+#                    * (0.5*(1+(frange/self.inj['f_break'])**(1/delta)))**((self.inj['alpha1']-self.inj['alpha2'])*delta) \
+#                    * 0.5 * (1 + np.tanh((self.inj['f_cut'] - frange)/self.inj['f_scale']))
+#            Omega_1mHz = (10**self.inj['log_omega0'])*(1e-3/self.inj['f_break'])**(self.inj['alpha1']) \
+#                    * (0.5*(1+(1e-3/self.inj['f_break'])**(1/delta)))**((self.inj['alpha1']-self.inj['alpha2'])*delta) \
+#                    * 0.5 * (1 + np.tanh((self.inj['f_cut'] - 1e-3)/self.inj['f_scale']))
+#        elif self.inj['spectral_inj'] == 'truncated_powerlaw':
+#            Omegaf = (10**self.inj['log_omega0'])*(frange/(self.params['fref']))**self.inj['alpha'] \
+#                    * 0.5 * (1 + np.tanh((self.inj['f_cut'] - frange)/self.inj['f_scale']))
+#            Omega_1mHz = 10**(self.inj['log_omega0']) * (1e-3/self.params['fref'])**(self.inj['alpha']) \
+#                    * 0.5 * (1 + np.tanh((self.inj['f_cut'] - 1e-3)/self.inj['f_scale']))
+#
+#        elif self.inj['spectral_inj'] == 'population':
+#            print("Constructing foreground spectrum from DWD population...")
+#            ## factor of two b/c (h_A,h_A*)~h^2~1/2 * S_A
+#            ## additional factor of 2 b/c S_GW = 2 * S_A
+#            Sgw = self.pop2spec(self.inj['popfile'],frange,self.params['dur']*u.s,names=self.inj['columns'],sep=self.inj['delimiter'])*4
+#            Omega_1mHz = Sgw[np.argmin(np.abs(frange - 1e-3))]/((3/(4*(1e-3)**3))*(H0/np.pi)**2)
+#        else:
+#            raise ValueError("Unknown spectral injection selected. Can be powerlaw, broken_powerlaw, free_broken_powerlaw, or population.")
+#        
+#        # Spectrum of the SGWB from Omegaf (population version goes directly to the spectrum from binary strains and frequencies)
+#        if self.inj['spectral_inj'] != 'population':
+#            Sgw = Omegaf*(3/(4*frange**3))*(H0/np.pi)**2    
+        N = self.Injection.Npersplice
+        halfN = int(0.5*N)
+        
+        injmodel_args = [injmodel.truevals[parameter] for parameter in injmodel.parameters]
+        
+        Sgw = injmodel.compute_Sgw(self.Injection.frange,injmodel_args)
+        
+        ## save frozen injected spectra
+        injmodel.frozen_spectra = Sgw
+        
         ## the spectrum of the frequecy domain gaussian for ifft
         norms = np.sqrt(self.params['fs']*Sgw*N)/2
 
@@ -548,11 +204,11 @@ class LISAdata(geometry, sph_geometry, instrNoise, populations):
             raise TypeError('Invalid specification of projection, projection can be E, G, or C')
 
         ## Loop over splice segments
-        for ii in range(nsplice):
+        for ii in range(self.Injection.nsplice):
 
             if self.inj['injtype'] == 'isgwb':
                 ## move frequency to be the zeroth-axis, then cholesky decomp
-                L_cholesky = norms[:, None, None] *  np.linalg.cholesky(np.moveaxis(response_mat[:, :, :, ii], -1, 0))
+                L_cholesky = norms[:, None, None] *  np.linalg.cholesky(np.moveaxis(injmodel.response_mat[:, :, :, ii], -1, 0))
 
             elif self.inj['injtype'] == 'sph_sgwb':
 
@@ -761,7 +417,7 @@ class LISAdata(geometry, sph_geometry, instrNoise, populations):
                     raise ValueError("Invalid specification of 'multi' type. Can be 'a' (anisotropic) or 'i' (isotropic).")
                 
             ## generate standard normal complex data frist
-            z_norm = np.random.normal(size=(frange.size, 3)) + 1j * np.random.normal(size=(frange.size, 3))
+            z_norm = np.random.normal(size=(self.Injection.frange.size, 3)) + 1j * np.random.normal(size=(self.Injection.frange.size, 3))
 
             ## The data in z_norm is rescaled into z_scale using L_cholesky
             z_scale = np.einsum('ijk, ikl -> ijl', L_cholesky, z_norm[:, :, None])[:, :, 0]
